Further API Cleanup and Improvement to ID Generators
Removed potentially harmful calls:
- Span.FinishWithTime
- Span.FinishWithDuration
- SpanOption: WithSpanContext()
- SpanOption: LocalEndpoint()

ID Generator SpanID() now takes a TraceID as its parameter. The Tracer will provide the actual TraceID for a new root span so the Generator can choose to use the TraceID for returning a root Span ID (saves cycles on edge nodes like API gateways). If the Tracer is creating a child span the TraceID provided to the SpanID() call will be empty.