feat: adds support for isNoop check on spans. (#181)
* feat: adds support for isNoop check on spans.

Span.IsNoop is specially usefull to avoid expensive operations when the tags or logs are not being recorded e.g. if you are about to read the body and process it to obtain certain tags, the check would save that computation as no changes will take effect.

* chore: fixes license year.

* chore: changes the IsNoop change into a package level function as per @basvanbeek's suggestion

IsNoop is more of the concern of the tracer and not necessarily of the span's concern, hence turning this method into a function.