add options to configure logger by interface
This commit provides users of this library with the ability to configure
any logger of their choosing to be passed into the various components of
this library. Doing so allows use cases such as using other open source
logging packages, e.g. logrus or zap, in conjunction with this library.

Closes https://github.com/openzipkin/zipkin-go/issues/37